Removed logic from calendar component, added standalone appointment service
